pyproject.toml
[project]
name = "sgx-filing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "beautifulsoup4>=4.14.2",
 "python-dotenv>=1.1.1",
 "rapidfuzz>=3.14.1",
 "requests>=2.32.5",
 "selenium>=4.36.0",
 "selenium-wire-2>=0.2.1",
 "supabase>=2.21.1",
 "typer>=0.19.2",
 "webdriver-manager>=4.0.2",
]
